% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_boem_orep_outlines.R
\docType{data}
\name{boem_orep_outlines}
\alias{boem_orep_outlines}
\alias{boem_wea_outlines}
\title{BOEM Office of Renewable Energy Program Lease and Wind Planning Areas}
\format{
An \code{sf} collection with 71 features and 30 fields.
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -158.6636 ymin: 20.7765 xmax: -68.5727 ymax: 43.9506}
\item{Geodetic CRS}{WGS 84}
}
}
\source{
\url{https://www.boem.gov/renewable-energy/mapping-and-data/renewable-energy-gis-data}
}
\usage{
data('boem_orep_outlines')
}
\description{
An \code{sf} object containing the outlines for BOEM Renewable Energy Lease (LEASE_STAGE = 'Active') and Wind Planning (LEASE_STAGE = 'Planning') Areas.
}
\details{
The Lease Area Boundaries were updated on 2025-08-05 and the Wind Planning Area Boundaries were updated on 2025-08-05. The Wind Planning Areas were rescinded on July 30, 2025.
}
\keyword{datasets}
