% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boemWind-package.R
\docType{package}
\name{boemWind-package}
\alias{boemWind}
\alias{boemWind-package}
\title{boemWind: BOEM Office of Renewable Energy Program GIS data for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide BOEM Office of Renewable Energy Program GIS data in sf objects stored in *.rda files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmhatch-noaa.github.io/boemWind/}
  \item Report bugs at \url{https://jmhatch-noaa.github.io/boemWind/issues}
}

}
\author{
\strong{Maintainer}: Joshua Hatch \email{joshua.hatch@noaa.gov}

}
\keyword{internal}
