% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boem_arcgis.R
\name{switch_url}
\alias{switch_url}
\title{switch_url}
\usage{
switch_url(type)
}
\arguments{
\item{type}{A character string for the type of GIS data that you would like to download.
Can be:
\itemize{
\item \code{active lease outlines}
\item \code{planning area outlines}
\item \code{project phases}
\item \code{export cables}
\item \code{substations}
\item \code{turbine locations}
\item \code{cable interconnections}
\item \code{data collection devices}
\item \code{cable corridors}
\item \code{cable landings}
\item \code{interarray cables}
}}
}
\value{
A url path.
}
\description{
Based on \code{type}, provide the Direct Service Link to download GIS data from BOEM's Renewable Energy Program using their ArcGIS server.
}
\keyword{internal}
