% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_usgs_offshore_wtbs.R
\docType{data}
\name{usgs_offshore_wtbs}
\alias{usgs_offshore_wtbs}
\title{The United States Wind Turbine Database (USWTDB)}
\format{
An \code{sf} collection with 7 features and 28 fields.
\describe{
\item{Geometry type}{POINT}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -75.4916 ymin: 36.8868 xmax: -71.5076 ymax: 41.1257}
\item{Geodetic CRS}{NAD83}
\item{Source}{data-raw/usgs-uswtdb-shapefile/uswtdb_v6_0_20230531.shp}
}
}
\source{
\url{https://eerscmap.usgs.gov/uswtdb/}
}
\usage{
data('usgs_offshore_wtbs')
}
\description{
An \code{sf} object containing the locations of offshore wind turbines.
}
\details{
Version: 6.0.0 - 2023-5-31
}
\keyword{datasets}
