% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_boem_wea_outlines.R
\docType{data}
\name{boem_wea_outlines}
\alias{boem_wea_outlines}
\title{BOEM Renewable Energy Lease Areas and Wind Planning Areas}
\format{
An \code{sf} collection with 74 features and 30 fields.
\describe{
\item{Geometry type}{MULTIPOLYGON}
\item{Dimension}{XY}
\item{Bounding box}{xmin: -158.6636 ymin: 20.7765 xmax: -68.3654 ymax: 43.9506}
\item{Geodetic CRS}{WGS 84}
}
}
\source{
\url{https://www.boem.gov/renewable-energy/mapping-and-data/renewable-energy-gis-data}
}
\usage{
data('boem_wea_outlines')
}
\description{
An \code{sf} object containing the outlines for BOEM Renewable Energy Lease Areas (LEASE_STAGE = 'Active') and Wind Planning Areas (LEASE_STAGE = 'Planning').
}
\details{
The Wind Lease Boundaries are from version 11.2 and the Wind Planning Area Boundaries are from version 11.2.
}
\keyword{datasets}
