% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boem_arcgis.R
\name{query_boem}
\alias{query_boem}
\title{Download GIS data from BOEM's Renewable Energy Program}
\usage{
query_boem(type)
}
\arguments{
\item{type}{A character string for the type of GIS data that you would like to download.
Can be:
\itemize{
\item \code{active lease outlines}
\item \code{planning area outlines}
\item \code{project phases}
\item \code{export cables}
\item \code{substations}
\item \code{turbine locations}
\item \code{cable interconnections}
\item \code{data collection devices}
\item \code{cable corridors}
\item \code{cable landings}
\item \code{interarray cables}
}}
}
\value{
A list with two named elements: 1) \code{meta} that stores metadata and 2) \code{data} that stores an \code{sf} data frame.
}
\description{
Query GIS data from BOEM's Renewable Energy Program using their ArcGIS server.
}
\examples{
# basic usage
active_leases <- query_boem('active lease outlines') |>
  purrr::pluck('data')

# use stored data instead
active_leases_v2 <- boem_wea_outlines |>
  dplyr::filter(LEASE_STAGE == 'Active')

}
